storage: move storage-client::sink to storage::sink::kafka::util
